reorg panther_box_helpers
