+++ rules/box_rules/box_suspicious_login_or_session.py
@@ -1,4 +1,5 @@
-from panther_base_helpers import box_parse_additional_details, deep_get
+from panther_base_helpers import deep_get
+from panther_box_helpers import box_parse_additional_details
 
 SUSPICIOUS_EVENT_TYPES = {
     "Suspicious Locations",
